Worker pool: handle tasks finished in other loops
Sometimes a task can finish while its event loop
is not active, if event loops are nested. Now we
make sure that these tasks are handled properly,
by checking for them at the start of the event loop
run.

We also need to make sure that the worker pool
is able to start tasks ASAP, even if they belong
to an event loop that is not active.