Fix growing promise chains
When a promise is replaced by another one.

Really replacing a promise with another one is tricky, because there
are parent and child references to promises. But it is done now.

It also breaks constructs that "remember" promises, because these
might be replaced by others. E.g. `when_all()` was such, and it
now has a new implementation.

A smaller issue was that `synchronise()` also stores a reference
to the top promise, which can be replaced, and then the reference
is not valid.

We probably could (and should) get rid of promise ids completely,
they don't seem to make much sense when promises can be replaced.
Right now we overwrite the id of replacing promise with the id
of the replaced promise, which is not great, but makes all tests
pass. It would be better not to use ids at all.

Closes #62.